+++ frontend/src/app/products/buy/buy.component.ts
@@ -15,14 +15,14 @@ import { TransactionsService } from '../transactions.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { stringify } from '@angular/compiler/src/util';
 import { MatDialog } from '@angular/material/dialog';
-import { OrderComponent } from 'src/app/order/order.component';
+import { OrderComponent } from '../order/order.component';
 
 @Component({
   selector: 'app-buy',
   templateUrl: './buy.component.html',
   styleUrls: ['./buy.component.css'],
 })
-export class BuyComponent extends OrderComponent implements OnInit  {
+export class BuyComponent implements OnInit {
   productId: string;
   product: IProduct;
   transactionPrice: number;
@@ -49,11 +49,8 @@ export class BuyComponent extends OrderComponent implements OnInit  {
     public router: Router,
     private productService: ProductsService,
     private transactionService: TransactionsService,
-    public dialog: MatDialog,
-    private orderComponent: OrderComponent
-  ) {
-    super(dialog, router)
-  }
+    public dialog: MatDialog //private orderComponent: OrderComponent
+  ) {}
 
   ngOnInit(): void {
     const productId = this.route.snapshot.paramMap.get('id');
@@ -62,58 +59,27 @@ export class BuyComponent extends OrderComponent implements OnInit  {
       this.product = product;
     });
   }
-  /*
-  calculatePrice(product: IProduct): void {
-    const rentalDays = this.rentalDaysForm.value;
-    console.log(rentalDays);
-    console.log(product.price);
-    const priceForCalc = product.price;
 
-    const totalPrice = (parseInt(rentalDays)*parseInt(priceForCalc))
-    const print = totalPrice;
-    console.log(print);
 
-    const price:number = product.price;
-    console.log(price);
-    console.log(isNaN(price));
-    console.log(isNaN(rentalDays));
-    console.log(typeof price);
-    console.log(typeof rentalDays);
-    const rentalDaysAsInt = parseInt(rentalDays);
-    console.log(typeof rentalDaysAsInt);
-    console.log( totalPrice )
-
-   //this.transactionPrice = (rentalDays)*(product.price);
-  // console.log(this.transactionPrice);
-  }
-
-   */
-
-  /*   get totalPrice(product){
-    return (product.price) * (pro.quantity) ;
-  } */
-
-  /* createDeliveryAddress(): void {
-    const deliveryAddress = this.deliveryForm.value;
-    console.log(deliveryAddress);
-  } */
-
-  confirmOrder(  ): void {
+  confirmOrder(): void {
     const order = {
       product: this.product,
       deliveryAddress: this.deliveryForm.value,
-      rentalDays: this.rentalDaysForm.value
-    }
-/*     product = this.product;
-    deliveryAddress = this.deliveryForm.value;
-    rentalDays = this.rentalDaysForm.value; */
-    this.orderComponent.openDialog(  );
+      rentalDays: this.rentalDaysForm.value,
+    };
 
-    //console.log("showing your order here: " + product.price + deliveryAddress.value + rentalDays.value)
 
-    // hier soll eine Zusammenfassung der Bestellung inkl der angegebenen Lieferadresse angezeigt werden
-    // über Button pay wird transaction erstellt
-    // mit zurück können Angaben geändert werden
+    this.dialog.open(OrderComponent, {
+      height: '400px',
+      width: '600px',
+      data: order,
+    }).afterClosed().subscribe(result => {
+      debugger
+      if (result) {
+        this.buy();
+      }
+    });
+
   }
 
   // User press button 'pay' which then comes here and runs pay/transaction
@@ -124,8 +90,8 @@ export class BuyComponent extends OrderComponent implements OnInit  {
     // order
     this.transactionService
       .pay(product, rentalDays, deliveryAddress)
-      .subscribe((res) => {});
+      .subscribe((res) => {
+        this.router.navigate(['products', 'history']);
+      });
   }
 }
-
-
