+++ frontend/src/app/products/products.module.ts
@@ -31,7 +31,7 @@ import {FilterComponent} from './filter/filter.component';
 import { ProductFilterPipe } from './pipes/product-filter.pipe';
 import { Ng2SearchPipeModule } from 'ng2-search-filter';
 import {MatBadgeModule} from '@angular/material/badge';
-import { OrderComponent } from '../order/order.component';
+import { OrderComponent } from './order/order.component';
 
 @NgModule({
   declarations: [
@@ -71,6 +71,9 @@ import { OrderComponent } from '../order/order.component';
   ],
   exports: [
     BrowseComponent
+  ],
+  entryComponents: [
+    OrderComponent
   ],
   providers: [
     ProductFilterPipe
